Add ability to start games offline and remove mock services
