+++ src/main/java/com/faforever/client/patch/SimpleHttpFeaturedModUpdater.java
@@ -1,13 +1,11 @@
 package com.faforever.client.patch;
 
-import com.faforever.client.FafClientApplication;
 import com.faforever.client.domain.FeaturedModBean;
 import com.faforever.client.task.TaskService;
 import lombok.RequiredArgsConstructor;
 import org.jetbrains.annotations.Nullable;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Lazy;
-import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
 import java.util.concurrent.CompletableFuture;
@@ -15,7 +13,6 @@ import java.util.concurrent.CompletableFuture;
 
 @Lazy
 @Component
-@Profile("!" + FafClientApplication.PROFILE_OFFLINE)
 @RequiredArgsConstructor
 public class SimpleHttpFeaturedModUpdater implements FeaturedModUpdater {
