--- src/main/java/com/faforever/client/remote/FafService.java
@@ -1,13 +1,0 @@
-package com.faforever.client.remote;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Lazy;
-import org.springframework.stereotype.Service;
-
-@Lazy
-@Service
-@RequiredArgsConstructor
-@Deprecated
-public class FafService {
-
-}
