+++ src/main/resources/i18n/messages.properties
@@ -243,6 +243,7 @@ login.forgot = Forgot login
 login.button.login = Login
 login.button.cancel = Cancel
 login.progress.message = Logging in…
+login.playOffline = Play Offline
 login.remember = Remember Me
 login.failed = Error occurred during login
 login.badState = State returned by user service does not match initial state if this continues to occur please reach out to technical help on the forum or discord channel
