+++ src/main/resources/theme/login/login.fxml
@@ -45,14 +45,15 @@
         <VBox fx:id="loginFormPane" alignment="TOP_CENTER" spacing="5.0" VBox.vgrow="SOMETIMES">
             <children>
                 <WebView fx:id="loginWebView" maxWidth="800.0" minHeight="100.0" minWidth="100.0" VBox.vgrow="ALWAYS"/>
-                <HBox alignment="CENTER" maxWidth="500" spacing="10.0" VBox.vgrow="NEVER">
+                <HBox alignment="CENTER" spacing="10.0" VBox.vgrow="NEVER">
                     <children>
                         <Label maxHeight="1.7976931348623157E308" minWidth="-Infinity" onMouseClicked="#onMouseClicked" text="%login.environment">
                             <HBox.margin>
                                 <Insets right="10.0" />
                             </HBox.margin>
                         </Label>
-                        <ComboBox fx:id="environmentComboBox" maxWidth="1.7976931348623157E308" HBox.hgrow="ALWAYS" />
+                        <ComboBox fx:id="environmentComboBox" maxWidth="1.7976931348623157E308" />
+                        <Button fx:id="playOfflineButton" text="%login.playOffline" onAction="#onPlayOfflineButtonClicked"/>
                         <CheckBox fx:id="rememberMeCheckBox" text="%login.remember" />
                     </children>
                 </HBox>
